Make -v/--verbose a global option
So far the -v/--verbose option was only accepted before any command or
sub commands were given. That is an unnecessary restriction that hinders
usability.
As it turns out, structopt actually support the notion of global
options, which can then be recognized as part of a command/sub command
as well. This change converts the -v/--verbose option into such a global
one.