Bump type_length_limit once more
With arbitrary code changes and/or changes to the tool chain we see
build breakages caused by some compiler internal type length limit being
exceeded. Because there is not much we can do other than bumping the
limit every time we see such a complaint, this change bumps it quite a
bit.