Fix total and base value calculation for position listing
We have noticed that the position listing is reporting inconsistent
values when we have short sell positions open.
Investigation has shown that the problem is that the cost basis reported
by Alpaca is actually negative for those positions. That screws up the
accumulation we do.
This change correct the issue by inverting the cost basis for short
positions.